Rename pydoctor.epydoc.markup.doctest to pydoctor.epydoc.doctest
Unlike the other modules in the 'markup' package, this module does not
implement a docstring format.
